requirements.txt
python-telegram-bot==20.7
requests==2.28.2
apscheduler==3.10.1
python-dotenv==1.0.0
beautifulsoup4==4.12.3